requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
behave==1.2.6
behave-django==1.4.0
coverage==5.2
Django==3.0.7
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==2.1.1
djangorestframework==3.11.0
importlib-metadata==1.6.1
parse==1.15.0
parse-type==0.5.2
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
